requirements.txt
openai
transformers
torch
haversine
pandas
fastapi
uvicorn[standard]
python-multipart
psutil
pygments
azure-storage-blob==12.19.0